+++ lib/src/frontend/throws_matcher.dart
@@ -8,6 +8,8 @@ import 'package:matcher/matcher.dart';
 
 import '../utils.dart';
 import 'async_matcher.dart';
+import '../frontend/test_chain.dart';
+import '../backend/invoker.dart';
 
 /// This function is deprecated.
 ///